+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fixing `#.edges(A, B)` with directed self loops.
 * Fixing case when `nodeAdded` events would not be fired when using `#.emptyCopy`.
 * Fixing `#.mergeEdgeWithKey` edge cases.
+* Dropping `before` and `after` from `replace` events metadata.
 * Correctly emitting `nodeAttributesUpdated` event when node is merged.
 * Correctly emitting `edgeAttributesUpdated` event when edge is merged.